transaction queue refactor: - removed blocking queue - removed loads of unnecessary semaphores - x64 support some x64 related fixes

+++ src/core/resource.h
@@ -35,7 +35,7 @@ namespace Lux
 
 		ObserverCallback& getObserverCb() { return m_cb; }
 
-		uint32_t size() const { return m_size; }
+		size_t size() const { return m_size; }
 
 		ResourceManager& getResourceManager() { return m_resource_manager; }
 
@@ -74,7 +74,7 @@ namespace Lux
 
 	protected:
 		Path m_path;
-		uint32_t m_size;
+		size_t m_size;
 		ObserverCallback m_cb;
 		ResourceManager& m_resource_manager;
 	};
